Templated alu testbench
To maximize code reusage, many blocks of the testbench Environment are
now generalized using type parameters when defining the classes.

alu_test works and found some errors in the dut when it comes to rotate
and shift functions: very likely it's due to the way these numeric_std
fuctions are implemented, since there are some warnings being printed by
the simulator that resemble overflows.